pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pythonclaw"
version = "0.2.0"
description = "Open-source autonomous AI agent framework — memory, RAG, skills, multi-channel, daemon, web dashboard."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Eric Wang", email = "wangchen2007915@gmail.com"},
]
keywords = ["llm", "agent", "ai", "autonomous", "memory", "rag", "skills", "telegram", "chatbot", "framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai",
]

[project.optional-dependencies]
all = [
    "anthropic",
    "google-generativeai",
    "python-telegram-bot[job-queue]>=20.0",
    "discord.py>=2.3",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "rank-bm25>=0.2",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tavily-python",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "beautifulsoup4",
    "requests",
]
anthropic = ["anthropic"]
gemini = ["google-generativeai"]
tavily = ["tavily-python"]
web = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]
telegram = [
    "python-telegram-bot[job-queue]>=20.0",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
]
discord = [
    "discord.py>=2.3",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
]
rag = [
    "rank-bm25>=0.2",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.scripts]
pythonclaw = "pythonclaw.main:main"

[project.urls]
Homepage = "https://github.com/ericwang915/PythonClaw"
Documentation = "https://github.com/ericwang915/PythonClaw#readme"
Repository = "https://github.com/ericwang915/PythonClaw"
Issues = "https://github.com/ericwang915/PythonClaw/issues"

[tool.setuptools.packages.find]
include = ["pythonclaw*"]

[tool.setuptools.package-data]
pythonclaw = [
    "templates/**/*",
    "web/static/**/*",
]
